+++ src/components/portfolio/sections/EducationSection.jsx
@@ -1,7 +1,7 @@
 import React, { forwardRef } from 'react';
-import './SectionStyles.css';
+import '../../sections/SectionStyles.css';
 
-const EducationSection = forwardRef(({ data, deviceType, isActive }, ref) => {
+const EducationSection = forwardRef(({ data, isActive }, ref) => {
 	if (!data) return null;
 
 	return (
